requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
dj-rest-auth==7.0.1
Django==5.0
django-cors-headers==4.6.0
django-drf-filepond==0.5.0
django-filter==24.3
django-storages==1.14.4
djangorestframework==3.15.2
idna==3.10
Markdown==3.7
mysqlclient==2.2.7
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
